+++ examples/babel-plugin-test/CHANGELOG.md
@@ -1,5 +1,12 @@
 # babel-plugin-test
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/babel-plugin-silk@2.0.1
+
 ## 0.0.2
 
 ### Patch Changes